--- src/components/Navbar/Navbar.module.css
@@ -1,20 +1,0 @@
-.navbar {
-    display: flex;
-    justify-content: space-between;
-    background-color: coral;
-}
-
-.logo {
-    padding: 10px;
-    align-self: center;
-}
-
-.buttons {
-    padding: 10px;
-    justify-content: center;
-}
-
-.logout {
-    padding: 10px;
-    border-radius: 0.5em;
-}
