+++ src/pages/Home/Home.js
@@ -1,5 +1,4 @@
 import styles from './Home.module.css';
-import Navbar from '../../components/Navbar/Navbar';
 import Chart from '../../components/Home/Chart';
 import Income from '../../components/Home/Income';
 import Goal from '../../components/Home/Goal';
@@ -61,7 +60,6 @@ const Home = ({ user }) => {
 
   return (
     <main className={styles.container}>
-      <Navbar/>
       <h1>hello, {user ? user.name : 'friend'}</h1>
       <h1>Today is {t}</h1>
       <Goal />
